backend - Copy/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.8.2
google-genai==0.8.0
playwright==1.44.0
python-telegram-bot==21.6
